citas Admin

+++ server/models/User.js
@@ -14,7 +14,7 @@ User.init(
       allowNull: false,
     },
     phone: {
-      type: S.INTEGER,
+      type: S.BIGINT,
       allowNull: false,
       validate:{
         min: 7,
